首页优化   睡觉

+++ MGLoveFreshBeen/MGLoveFreshBeen/Class/Home/C/HomeVC.m
@@ -184,7 +184,7 @@ static NSString *const KHomeFooterIdentifier = @"Footer";
     // hotView点击
     [MGNotificationCenter addObserverForName:MGHotPanClickNotification object:nil queue:nil usingBlock:^(NSNotification * note) {
         
-        int item = [[note.userInfo valueForKeyPath:@"tag"] integerValue] - 20;
+        NSInteger item = ([[note.userInfo valueForKeyPath:@"tag"] integerValue] - 20);
         
          HomeWebVC *webVC = [[HomeWebVC alloc] initWithNavigationTitle:[_headData.data.icons[item] valueForKeyPath:@"name"] withUrlStr:[_headData.data.icons[item] valueForKeyPath:@"customURL"]];
             [weakSelf.navigationController pushViewController:webVC animated:YES];
